tests/requirements.txt
anyio==4.3.0
aiokafka==0.9.0
alembic==1.11.1
Authlib==1.2.0
clickhouse-driver==0.2.6
fastapi==0.105.0
grpcio==1.62.1
grpcio-tools==1.62.0
grpclib==0.4.6
httpx==0.24.1
ipython==8.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
asyncpg==0.28.0
pytest==7.3.2
requests==2.31.0
SQLAlchemy==2.0.16
uvicorn==0.22.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy-utils==0.41.1
python-multipart==0.0.9
pyjwt==2.7.0
python-dateutil==2.8.2
pydantic-settings==2.0.2
pydantic[email]
lxml==4.9.3
python-jose==3.3.0
# dev
black==23.3.0
flake8==6.0.0
isort==5.12.0
mypy==1.4.0
types-passlib==1.7.7.12
expects==0.9.0
types-requests==2.31.0.7
pytest==8.1.2
pytest-asyncio==0.23.6
testcontainers==4.5.0